Added errors, control exceptions and unitary test for the controllers

+++ src/main/java/challenge/services/PokemonServiceImpl.java
@@ -1,19 +1,28 @@
 package challenge.services;
 
 import challenge.entities.Pokemon;
+import challenge.exception.types.ChallengeServiceException;
 import challenge.repositories.PokemonRepository;
 import challenge.search.PokemonSearch;
+import challenge.utils.ErrorCodes;
+import challenge.utils.ErrorMessages;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@EnableConfigurationProperties
 public class PokemonServiceImpl implements PokemonService {
 
     @Autowired
     PokemonRepository pokemonRepository;
+
+    @Autowired
+    ErrorMessages errorMessages;
 
     @Override
     public Pokemon findPokemonById(String id) {
@@ -28,16 +37,34 @@ public class PokemonServiceImpl implements PokemonService {
 
     @Override
     public Pokemon savePokemon(Pokemon pokemon) {
-        return (Pokemon) this.pokemonRepository.save(pokemon);
+        try {
+            pokemon.setId(null);
+            return (Pokemon) this.pokemonRepository.save(pokemon);
+        } catch (DuplicateKeyException e) {
+            throw new ChallengeServiceException(errorMessages.getProperty(ErrorCodes.DUPLICATE_POKEMON));
+        }
     }
 
     @Override
     public Pokemon updatePokemon(Pokemon pokemon) {
-        return (Pokemon) this.pokemonRepository.save(pokemon);
+        try {
+            this.pokemonRepository.updatePokemon(pokemon);
+            return this.findPokemonById(pokemon.getId());
+        } catch (Exception e) {
+            throw new ChallengeServiceException(errorMessages.getProperty(ErrorCodes.UPDATE_ERROR));
+        }
     }
 
     @Override
     public void deletePokemon(String id) {
-        this.pokemonRepository.delete(id);
+        Pokemon pokemon = this.findPokemonById(id);
+        if (null == pokemon) {
+            throw new ChallengeServiceException(errorMessages.getProperty(ErrorCodes.POKEMON_NOT_FOUND));
+        }
+        try {
+            this.pokemonRepository.delete(id);
+        } catch (Exception e) {
+            throw new ChallengeServiceException(errorMessages.getProperty(ErrorCodes.DELETE_ERROR));
+        }
     }
 }
